+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [*Unreleased*]
+
+### Changed
+
+* Changed signature for `onSubmitAction` to use new `ISubmitActionRequest` interface
+
 ## [*1.2.1*] - <*2019-05-07*>
 
 ### Changed